src/Makefile.am: avoid use of GNU-specific form of $<
$< is only allowed in implicit rules in some non-GNU makes (e.g. BSD,
AIX) so avoid this use by referencing the dependent curl.1 file directly
instead. This is somewhat tricky because the file is supplied in the
packaged tar ball (but not in git) but must still be able to be rebuilt
when its dependencies change. The right thing must happen in both tar
ball and git source trees, as well as in in-tree and out-of-tree builds.